# openAPI_Spec/admin端接口.openapi.yaml
openapi: 3.1.0
info:
  title: admin端接口
  description: ''
  version: 1.0.0
...
paths:
...
  /admin/employee/editPassword:
    put:
      summary: 修改密码
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                empId:
                  type: integer
                  format: int64
                  description: 员工id
                newPassword:
                  type: string
                  description: 新密码
                oldPassword:
                  type: string
                  description: 旧密码
              title: PasswordEditDTO
              $$ref: '#/definitions/PasswordEditDTO'
              required:
                - empId
                - newPassword
                - oldPassword
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/employee/status/{status}:
    post:
      summary: 启用、禁用员工账号
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: status
          in: path
          description: 状态，1为启用 0为禁用
          required: true
          example: '1'
          schema:
            type: string
        - name: id
          in: query
          description: 员工id
          required: true
          example: ''
          schema:
            type: string
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/employee/page:
    get:
      summary: 员工分页查询
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: name
          in: query
          description: 员工姓名
          required: false
          example: 张三
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          required: true
          example: '1'
          schema:
            type: string
        - name: pageSize
          in: query
          description: 每页记录数
          required: true
          example: '10'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                  msg:
                    type: 'null'
                  data:
                    type: object
                    properties:
                      total:
                        type: number
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: number
                            username:
                              type: string
                            name:
                              type: string
                            password:
                              type: string
                            phone:
                              type: string
                            sex:
                              type: string
                            idNumber:
                              type: string
                            status:
                              type: number
                            createTime:
                              type:
                                - string
                                - 'null'
                            updateTime:
                              type: string
                            createUser:
                              type:
                                - number
                                - 'null'
                            updateUser:
                              type: number
                          required:
                            - id
                            - username
                            - name
                            - password
                            - phone
                            - sex
                            - idNumber
                            - status
                            - createTime
                            - updateTime
                            - createUser
                            - updateUser
                    required:
                      - total
                      - records
                required:
                  - code
                  - data
          headers: {}
      security: []
  /admin/employee/login:
    post:
      summary: 员工登录
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: 密码
                username:
                  type: string
                  description: 用户名
              title: EmployeeLoginDTO
              description: 员工登录时传递的数据模型
              $$ref: '#/definitions/EmployeeLoginDTO'
              required:
                - password
                - username
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                        description: 主键值
                      name:
                        type: string
                        description: 姓名
                      token:
                        type: string
                        description: jwt令牌
                      userName:
                        type: string
                        description: 用户名
                    title: EmployeeLoginVO
                    description: 员工登录返回的数据格式
                    $$ref: '#/definitions/EmployeeLoginVO'
                  msg:
                    type: string
                title: R«EmployeeLoginVO»
                $$ref: '#/definitions/R«EmployeeLoginVO»'
                required:
                  - code
          headers: {}
      security: []
  /admin/employee:
    post:
      summary: 新增员工
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  description: 员工id
                idNumber:
                  type: string
                  description: 身份证
                name:
                  type: string
                  description: 姓名
                phone:
                  type: string
                  description: 手机号
                sex:
                  type: string
                  description: 性别
                username:
                  type: string
                  description: 用户名
              title: EmployeeDTO
              $$ref: '#/definitions/EmployeeDTO'
              required:
                - idNumber
                - name
                - phone
                - sex
                - username
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties: {}
                  msg:
                    type: string
                title: R
                $$ref: '#/definitions/R'
                required:
                  - code
          headers: {}
      security: []
    put:
      summary: 编辑员工信息
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                idNumber:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                sex:
                  type: string
                username:
                  type: string
              title: EmployeeDTO
              $$ref: '#/definitions/EmployeeDTO'
              required:
                - id
                - idNumber
                - name
                - phone
                - sex
                - username
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/employee/{id}:
    get:
      summary: 根据id查询员工
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: id
          in: path
          description: 员工id
          required: true
          example: '100'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      createTime:
                        type: string
                        format: date-time
                      createUser:
                        type: integer
                        format: int64
                      id:
                        type: integer
                        format: int64
                      idNumber:
                        type: string
                      name:
                        type: string
                      password:
                        type: string
                      phone:
                        type: string
                      sex:
                        type: string
                      status:
                        type: integer
                        format: int32
                      updateTime:
                        type: string
                        format: date-time
                      updateUser:
                        type: integer
                        format: int64
                      username:
                        type: string
                    title: Employee
                    $$ref: '#/definitions/Employee'
                  msg:
                    type: string
                title: R«Employee»
                $$ref: '#/definitions/R«Employee»'
                required:
                  - code
                  - data
          headers: {}
      security: []
  /admin/employee/logout:
    post:
      summary: 退出登录
      deprecated: false
      description: ''
      tags:
        - 员工相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: empty object
              properties: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []